Refactor text rendering to use instanced rendering
Now most of the work is done in the vertex shader.
Instead of a massive vertex buffer with all characters,
we now use instancing to render each character with its own glyph data.
Also prepopulate glyph attributes in numpy arrays so that we can build the instance buffers using vectorized operations instead of slow Python loops